split up to different files the rest client and auth REST API calls
